third_party/WebKit/LayoutTests/platform/mac-retina/external/wpt/battery-status/battery-full-manual.https-expected.txt

This is a testharness.js-based test.
FAIL Battery Test: battery full, charger plugged in assert_equals: chargingTime must be set to 0 expected 0 but got Infinity
Harness: the test ran to completion.
